+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/common/properties/SubmodelProperties.java
@@ -1,11 +1,8 @@
 package org.eclipse.tractusx.traceability.common.properties;
 
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.stereotype.Component;
 
 @Configuration
 @Data
@@ -18,4 +15,5 @@ public class SubmodelProperties {
     private String clientSecret;
     private String tokenUrl;
     private String oauthProviderRegistrationId;
+    private Boolean useCustomImplementation;
 }